+++ exchange-v2/contracts/ExchangeV2.sol
@@ -20,4 +20,5 @@ contract ExchangeV2 is ExchangeV2Core, RaribleTransferManager {
         __RaribleTransferManager_init_unchained(newProtocolFee, newDefaultFeeReceiver, newRoyaltiesProvider);
         __OrderValidator_init_unchained();
     }
+
 }