+++ transfer-manager/contracts/RaribleTransferManager.sol
@@ -18,22 +18,41 @@ abstract contract RaribleTransferManager is OwnableUpgradeable, ITransferManager
     using BpLibrary for uint;
     using SafeMathUpgradeable for uint;
 
-    // @notice protocolFee is deprecated 
-    uint private protocolFee;
+    ProtocolFeeData public protocolFee;
     IRoyaltiesProvider public royaltiesRegistry;
 
-    // deprecated: no need without protocolFee
+    //deprecated
     address private defaultFeeReceiver;
-    // deprecated: no need without protocolFee 
+    // deprecated
     mapping(address => address) private feeReceivers;
 
+    /// @dev event that's emitted when ProtocolFeeData buyerAmount changes
+    event BuyerFeeAmountChanged(uint oldValue, uint newValue);
+
+    /// @dev event that's emitted when ProtocolFeeData sellerAmount changes
+    event SellerFeeAmountChanged(uint oldValue, uint newValue);
+
+    /// @dev event that's emitted when ProtocolFeeData receiver changes
+    event FeeReceiverChanged(address oldValue, address newValue);
+
+    /// @dev struct to store protocol fee - receiver address, buyer fee amount (in bp), seller fee amount (in bp)
+    struct ProtocolFeeData {
+        address receiver;
+        uint48 buyerAmount;
+        uint48 sellerAmount;
+    }
+
+    /**
+        @notice initialises RaribleTransferManager state
+        @param newProtocolFee deprecated
+        @param newDefaultFeeReceiver deprecated
+        @param newRoyaltiesProvider royaltiesRegistry contract address
+     */
     function __RaribleTransferManager_init_unchained(
         uint newProtocolFee,
         address newDefaultFeeReceiver,
         IRoyaltiesProvider newRoyaltiesProvider
     ) internal initializer {
-        protocolFee = newProtocolFee;
-        defaultFeeReceiver = newDefaultFeeReceiver;
         royaltiesRegistry = newRoyaltiesProvider;
     }
 
@@ -41,27 +60,48 @@ abstract contract RaribleTransferManager is OwnableUpgradeable, ITransferManager
         royaltiesRegistry = newRoyaltiesRegistry;
     }
 
+    function setPrtocolFeeReceiver(address _receiver) public onlyOwner {
+        emit FeeReceiverChanged(protocolFee.receiver, _receiver);
+        protocolFee.receiver = _receiver;
+    }
+
+    function setPrtocolFeeBuyerAmount(uint48 _buyerAmount) public onlyOwner {
+        emit BuyerFeeAmountChanged(protocolFee.buyerAmount, _buyerAmount);
+        protocolFee.buyerAmount = _buyerAmount;
+    }
+
+    function setPrtocolFeeSellerAmount(uint48 _sellerAmount) public onlyOwner {
+        emit SellerFeeAmountChanged(protocolFee.sellerAmount, _sellerAmount);
+        protocolFee.sellerAmount = _sellerAmount;
+    }
+
+    function setAllProtocolFeeData(address _receiver, uint48 _buyerAmount, uint48 _sellerAmount) public onlyOwner {
+        setPrtocolFeeReceiver(_receiver);
+        setPrtocolFeeBuyerAmount(_buyerAmount);
+        setPrtocolFeeSellerAmount(_sellerAmount);
+    }
+
     /**
         @notice executes transfers for 2 matched orders
         @param left DealSide from the left order (see LibDeal.sol)
         @param right DealSide from the right order (see LibDeal.sol)
-        @param dealData DealData of the match (see LibDeal.sol)
+        @param feeSide feeSide of the match
         @return totalLeftValue - total amount for the left order
         @return totalRightValue - total amout for the right order
     */
     function doTransfers(
         LibDeal.DealSide memory left,
         LibDeal.DealSide memory right,
-        LibDeal.DealData memory dealData
+        LibFeeSide.FeeSide feeSide
     ) override internal returns (uint totalLeftValue, uint totalRightValue) {
         totalLeftValue = left.asset.value;
         totalRightValue = right.asset.value;
 
-        if (dealData.feeSide == LibFeeSide.FeeSide.LEFT) {
-            totalLeftValue = doTransfersWithFees(left, right, dealData.maxFeesBasePoint);
+        if (feeSide == LibFeeSide.FeeSide.LEFT) {
+            totalLeftValue = doTransfersWithFees(left, right, protocolFee);
             transferPayouts(right.asset.assetType, right.asset.value, right.from, left.payouts, right.proxy);
-        } else if (dealData.feeSide == LibFeeSide.FeeSide.RIGHT) {
-            totalRightValue = doTransfersWithFees(right, left, dealData.maxFeesBasePoint);
+        } else if (feeSide == LibFeeSide.FeeSide.RIGHT) {
+            totalRightValue = doTransfersWithFees(right, left,protocolFee);
             transferPayouts(left.asset.assetType, left.asset.value, left.from, right.payouts, left.proxy);
         } else {
             transferPayouts(left.asset.assetType, left.asset.value, left.from, right.payouts, left.proxy);
@@ -73,16 +113,16 @@ abstract contract RaribleTransferManager is OwnableUpgradeable, ITransferManager
         @notice executes the fee-side transfers (payment + fees)
         @param paymentSide DealSide of the fee-side order
         @param nftSide  DealSide of the nft-side order
-        @param maxFeesBasePoint max fee for the sell-order (used and is > 0 for V3 orders only)
+        @param _protocolFee protocol fee data
         @return totalAmount of fee-side asset
     */
     function doTransfersWithFees(
         LibDeal.DealSide memory paymentSide,
         LibDeal.DealSide memory nftSide,
-        uint maxFeesBasePoint
+        ProtocolFeeData memory _protocolFee
     ) internal returns (uint totalAmount) {
-        totalAmount = calculateTotalAmount(paymentSide.asset.value, paymentSide.originFees, maxFeesBasePoint);
-        uint rest = totalAmount;
+        totalAmount = calculateTotalAmount(paymentSide.asset.value, _protocolFee, paymentSide.originFees);
+        uint rest = transferProtocolFee(totalAmount, paymentSide.asset.value, paymentSide.from, _protocolFee, paymentSide.asset.assetType, paymentSide.proxy);
 
         rest = transferRoyalties(paymentSide.asset.assetType, nftSide.asset.assetType, nftSide.payouts, rest, paymentSide.asset.value, paymentSide.from, paymentSide.proxy);
         if (
@@ -99,6 +139,21 @@ abstract contract RaribleTransferManager is OwnableUpgradeable, ITransferManager
             (rest,) = transferFees(paymentSide.asset.assetType, rest, paymentSide.asset.value, nftSide.originFees, paymentSide.from, paymentSide.proxy);
         }
         transferPayouts(paymentSide.asset.assetType, rest, paymentSide.from, nftSide.payouts, paymentSide.proxy);
+    }
+
+    function transferProtocolFee(
+        uint totalAmount,
+        uint amount,
+        address from,
+        ProtocolFeeData memory _protocolFee,
+        LibAsset.AssetType memory matchCalculate,
+        address proxy
+    ) internal returns (uint) {
+        (uint rest, uint fee) = subFeeInBp(totalAmount, amount, _protocolFee.buyerAmount + _protocolFee.sellerAmount);
+        if (fee > 0) {
+            transfer(LibAsset.Asset(matchCalculate, fee), from, _protocolFee.receiver, proxy);
+        }
+        return rest;
     }
 
     /**
@@ -223,23 +278,22 @@ abstract contract RaribleTransferManager is OwnableUpgradeable, ITransferManager
     /**
         @notice calculates total amount of fee-side asset that is going to be used in match
         @param amount fee-side order value
+        @param _protocolFee protocol fee
         @param orderOriginFees fee-side order's origin fee (it adds on top of the amount)
-        @param maxFeesBasePoint max fee for the sell-order (used and is > 0 for V3 orders only)
         @return total amount of fee-side asset
     */
     function calculateTotalAmount(
         uint amount,
-        LibPart.Part[] memory orderOriginFees,
-        uint maxFeesBasePoint
+        ProtocolFeeData memory _protocolFee,
+        LibPart.Part[] memory orderOriginFees
     ) internal pure returns (uint) {
-        if (maxFeesBasePoint > 0) {
-            return amount;
-        }
-        uint fees = 0;
+        
+        uint fees = _protocolFee.buyerAmount;
         for (uint256 i = 0; i < orderOriginFees.length; ++i) {
             require(orderOriginFees[i].value <= 10000, "origin fee is too big");
             fees = fees + orderOriginFees[i].value;
         }
+
         return amount.add(amount.bp(fees));
     }