+++ CHANGELOG.md
@@ -3,6 +3,113 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.1](https://github.com/rarible/protocol-contracts/compare/v0.7.15...v0.8.1) (2023-11-20)
+
+
+### Bug Fixes
+
+* add gitignore ([cb22f28](https://github.com/rarible/protocol-contracts/commit/cb22f289ca98e233f86689a259e1092241d0077e))
+* cleanup ([7113d79](https://github.com/rarible/protocol-contracts/commit/7113d7921da1a1dfec4883d73014a012d6ad01ef))
+* connection ([25e0320](https://github.com/rarible/protocol-contracts/commit/25e032015de29427a79f95d9a475885f26cc8ba9))
+* hd-provider ([444cb42](https://github.com/rarible/protocol-contracts/commit/444cb421127fe787a3678733e46425f733e65e8b))
+* remove build ([643ecbb](https://github.com/rarible/protocol-contracts/commit/643ecbbf25ca4f81ec47e2f7d3c9a6dd2b95c0b3))
+* remove generated files ([13f6ee7](https://github.com/rarible/protocol-contracts/commit/13f6ee72deab87f92c65114579d33fbfc0472de4))
+
+
+### Features
+
+* add arbitrum_mainnet config ([ba3d04a](https://github.com/rarible/protocol-contracts/commit/ba3d04aec3c5a91f600b279e39b5844e847e1566))
+* add build ([cd66e8e](https://github.com/rarible/protocol-contracts/commit/cd66e8e7d0919076eb5cfc778083fdf3e17cdf64))
+* add chilis mainnet config ([dae6108](https://github.com/rarible/protocol-contracts/commit/dae6108373e592fa3bd3bbffe4209b508e10c2a1))
+* add gas multiplier for the arbitrum_seploia ([15fbab6](https://github.com/rarible/protocol-contracts/commit/15fbab6a584bfeb0ba465d4b85ffe63f56fa9bb0))
+* add hardhat-deploy package ([704d1e1](https://github.com/rarible/protocol-contracts/commit/704d1e15fad6026bf25d0fd2319f3b0254b5c15f))
+* add latest changes ([b4b454c](https://github.com/rarible/protocol-contracts/commit/b4b454c1f7e2ea2d86d3babe3e0294ea6d7baefd))
+* add missing deployment ([93b73f1](https://github.com/rarible/protocol-contracts/commit/93b73f184546ccf366a2f330bf12521649745a4c))
+* add missing deployment ([041bb69](https://github.com/rarible/protocol-contracts/commit/041bb690294ec14f29373174c1ee25ec86ecab7b))
+* add new package in package.json ([9f6f7f0](https://github.com/rarible/protocol-contracts/commit/9f6f7f05fdf5ddef49a084302d81dff224e5ab6e))
+* add verification for arbitrum_sepolia ([abfefd9](https://github.com/rarible/protocol-contracts/commit/abfefd9bb95a0055a0c8386ef59072d8f9212382))
+* add verify ([1644fbd](https://github.com/rarible/protocol-contracts/commit/1644fbd31d2a0c2593cb874373af4734df43ad06))
+* add zkatana testnet deploy ([b775102](https://github.com/rarible/protocol-contracts/commit/b775102209edc789e5a190438215b557c20cc421))
+* deploy contracts on arbitrum_sepolia ([2c45266](https://github.com/rarible/protocol-contracts/commit/2c45266ee2067229045584a19a539232db46eb0b))
+* deployed on arbitrum (except tokens) ([11efcf2](https://github.com/rarible/protocol-contracts/commit/11efcf24de934f358222d91a73cd1ba758c5fd36))
+* deployed part of the contracts to arbitrum mainnet ([df0b102](https://github.com/rarible/protocol-contracts/commit/df0b1022e3c04de342bebeba1ede2fde50ac0a50))
+* deployment to chiliz ([6e012a7](https://github.com/rarible/protocol-contracts/commit/6e012a713804302775da782380038019984e9013))
+* fix verify ([ffa748b](https://github.com/rarible/protocol-contracts/commit/ffa748b2364fb279fd586ffc4d1183aa4f99be03))
+* fix verify settings ([9ac59a7](https://github.com/rarible/protocol-contracts/commit/9ac59a7621f272b43a8458413f83faad7bed52b6))
+* packages sync ([58487ea](https://github.com/rarible/protocol-contracts/commit/58487ea1c48cea8ebea68185c0e612ffd31bea2f))
+* plugin verify ([b2c7262](https://github.com/rarible/protocol-contracts/commit/b2c7262daaaed9fad799203932cee1ee22d4fc5d))
+* PT-3744 artblocksV2 royalties provider ([06ea3b3](https://github.com/rarible/protocol-contracts/commit/06ea3b335009a6f7006199c077f17a0b3141c989))
+* PT-4067-undeprecated protocol fee, deprecated V3 orders ([747e585](https://github.com/rarible/protocol-contracts/commit/747e585ff9dad2411af83014c4f5c9adbe2d8d61))
+* PT-4236 make hardhat-deploy/ package with migrations in hardhat. only necessary migrations: RoyaltiesRegistry, ExchangeV2,  4 TransferProxies, Collection asset matcher, rarible collections + factories. supports configurations to deploy contracts with supported meta transaction or not. added integration tests + verification after deploy. ([a28af59](https://github.com/rarible/protocol-contracts/commit/a28af5965f398d9c75eb38041c2d1dbcc5651cdd))
+* remove build ([f572a98](https://github.com/rarible/protocol-contracts/commit/f572a98439288856ea303b8eaf835ba9e159dc82))
+* remove types ([26db2ce](https://github.com/rarible/protocol-contracts/commit/26db2cec680d566d4f12a3a56991a56ba895343e))
+* set version ([419b05a](https://github.com/rarible/protocol-contracts/commit/419b05a732e611b67d2a598f0f6d8a20231e1c64))
+* un-deprecate protocol fee (one side) ([2cc0b3a](https://github.com/rarible/protocol-contracts/commit/2cc0b3a086ebcf2ccdba34e9fa67158ae6d44532))
+* undeprecate protocol fee ([9c589a5](https://github.com/rarible/protocol-contracts/commit/9c589a57028b2f541245f0e96557c535d1740bf9))
+
+
+### Reverts
+
+* Revert "feat: PT-3721, PT-3789 support OFR in tokens" ([f42da29](https://github.com/rarible/protocol-contracts/commit/f42da29424aa03b2af6e9bb62df734395b66ba00))
+* Revert "chore: update tokens on goerli" ([9a0071d](https://github.com/rarible/protocol-contracts/commit/9a0071d9361b2315dc0dd822cee401689e23fe2e))
+
+
+
+
+
+# [0.8.0](https://github.com/rarible/protocol-contracts/compare/v0.7.15...v0.8.0) (2023-11-20)
+
+
+### Bug Fixes
+
+* add gitignore ([cb22f28](https://github.com/rarible/protocol-contracts/commit/cb22f289ca98e233f86689a259e1092241d0077e))
+* cleanup ([7113d79](https://github.com/rarible/protocol-contracts/commit/7113d7921da1a1dfec4883d73014a012d6ad01ef))
+* connection ([25e0320](https://github.com/rarible/protocol-contracts/commit/25e032015de29427a79f95d9a475885f26cc8ba9))
+* hd-provider ([444cb42](https://github.com/rarible/protocol-contracts/commit/444cb421127fe787a3678733e46425f733e65e8b))
+* remove build ([643ecbb](https://github.com/rarible/protocol-contracts/commit/643ecbbf25ca4f81ec47e2f7d3c9a6dd2b95c0b3))
+* remove generated files ([13f6ee7](https://github.com/rarible/protocol-contracts/commit/13f6ee72deab87f92c65114579d33fbfc0472de4))
+
+
+### Features
+
+* add arbitrum_mainnet config ([ba3d04a](https://github.com/rarible/protocol-contracts/commit/ba3d04aec3c5a91f600b279e39b5844e847e1566))
+* add build ([cd66e8e](https://github.com/rarible/protocol-contracts/commit/cd66e8e7d0919076eb5cfc778083fdf3e17cdf64))
+* add chilis mainnet config ([dae6108](https://github.com/rarible/protocol-contracts/commit/dae6108373e592fa3bd3bbffe4209b508e10c2a1))
+* add gas multiplier for the arbitrum_seploia ([15fbab6](https://github.com/rarible/protocol-contracts/commit/15fbab6a584bfeb0ba465d4b85ffe63f56fa9bb0))
+* add hardhat-deploy package ([704d1e1](https://github.com/rarible/protocol-contracts/commit/704d1e15fad6026bf25d0fd2319f3b0254b5c15f))
+* add latest changes ([b4b454c](https://github.com/rarible/protocol-contracts/commit/b4b454c1f7e2ea2d86d3babe3e0294ea6d7baefd))
+* add missing deployment ([93b73f1](https://github.com/rarible/protocol-contracts/commit/93b73f184546ccf366a2f330bf12521649745a4c))
+* add missing deployment ([041bb69](https://github.com/rarible/protocol-contracts/commit/041bb690294ec14f29373174c1ee25ec86ecab7b))
+* add new package in package.json ([9f6f7f0](https://github.com/rarible/protocol-contracts/commit/9f6f7f05fdf5ddef49a084302d81dff224e5ab6e))
+* add verification for arbitrum_sepolia ([abfefd9](https://github.com/rarible/protocol-contracts/commit/abfefd9bb95a0055a0c8386ef59072d8f9212382))
+* add verify ([1644fbd](https://github.com/rarible/protocol-contracts/commit/1644fbd31d2a0c2593cb874373af4734df43ad06))
+* add zkatana testnet deploy ([b775102](https://github.com/rarible/protocol-contracts/commit/b775102209edc789e5a190438215b557c20cc421))
+* deploy contracts on arbitrum_sepolia ([2c45266](https://github.com/rarible/protocol-contracts/commit/2c45266ee2067229045584a19a539232db46eb0b))
+* deployed on arbitrum (except tokens) ([11efcf2](https://github.com/rarible/protocol-contracts/commit/11efcf24de934f358222d91a73cd1ba758c5fd36))
+* deployed part of the contracts to arbitrum mainnet ([df0b102](https://github.com/rarible/protocol-contracts/commit/df0b1022e3c04de342bebeba1ede2fde50ac0a50))
+* deployment to chiliz ([6e012a7](https://github.com/rarible/protocol-contracts/commit/6e012a713804302775da782380038019984e9013))
+* fix verify ([ffa748b](https://github.com/rarible/protocol-contracts/commit/ffa748b2364fb279fd586ffc4d1183aa4f99be03))
+* fix verify settings ([9ac59a7](https://github.com/rarible/protocol-contracts/commit/9ac59a7621f272b43a8458413f83faad7bed52b6))
+* packages sync ([58487ea](https://github.com/rarible/protocol-contracts/commit/58487ea1c48cea8ebea68185c0e612ffd31bea2f))
+* plugin verify ([b2c7262](https://github.com/rarible/protocol-contracts/commit/b2c7262daaaed9fad799203932cee1ee22d4fc5d))
+* PT-3744 artblocksV2 royalties provider ([06ea3b3](https://github.com/rarible/protocol-contracts/commit/06ea3b335009a6f7006199c077f17a0b3141c989))
+* PT-4067-undeprecated protocol fee, deprecated V3 orders ([747e585](https://github.com/rarible/protocol-contracts/commit/747e585ff9dad2411af83014c4f5c9adbe2d8d61))
+* PT-4236 make hardhat-deploy/ package with migrations in hardhat. only necessary migrations: RoyaltiesRegistry, ExchangeV2,  4 TransferProxies, Collection asset matcher, rarible collections + factories. supports configurations to deploy contracts with supported meta transaction or not. added integration tests + verification after deploy. ([a28af59](https://github.com/rarible/protocol-contracts/commit/a28af5965f398d9c75eb38041c2d1dbcc5651cdd))
+* remove build ([f572a98](https://github.com/rarible/protocol-contracts/commit/f572a98439288856ea303b8eaf835ba9e159dc82))
+* remove types ([26db2ce](https://github.com/rarible/protocol-contracts/commit/26db2cec680d566d4f12a3a56991a56ba895343e))
+* set version ([419b05a](https://github.com/rarible/protocol-contracts/commit/419b05a732e611b67d2a598f0f6d8a20231e1c64))
+* un-deprecate protocol fee (one side) ([2cc0b3a](https://github.com/rarible/protocol-contracts/commit/2cc0b3a086ebcf2ccdba34e9fa67158ae6d44532))
+
+
+### Reverts
+
+* Revert "feat: PT-3721, PT-3789 support OFR in tokens" ([f42da29](https://github.com/rarible/protocol-contracts/commit/f42da29424aa03b2af6e9bb62df734395b66ba00))
+* Revert "chore: update tokens on goerli" ([9a0071d](https://github.com/rarible/protocol-contracts/commit/9a0071d9361b2315dc0dd822cee401689e23fe2e))
+
+
+
+
+
 ## [0.7.15](https://github.com/rarible/protocol-contracts/compare/v0.7.14-beta.3...v0.7.15) (2023-10-10)
 
 **Note:** Version bump only for package root