+++ interfaces/CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.1](https://github.com/rariblecom/protocol-contracts/compare/v0.7.15...v0.8.1) (2023-11-20)
+
+
+### Bug Fixes
+
+* add gitignore ([cb22f28](https://github.com/rariblecom/protocol-contracts/commit/cb22f289ca98e233f86689a259e1092241d0077e))
+* remove generated files ([13f6ee7](https://github.com/rariblecom/protocol-contracts/commit/13f6ee72deab87f92c65114579d33fbfc0472de4))
+
+
+### Features
+
+* undeprecate protocol fee ([9c589a5](https://github.com/rariblecom/protocol-contracts/commit/9c589a57028b2f541245f0e96557c535d1740bf9))
+
+
+
+
+
+# [0.8.0](https://github.com/rariblecom/protocol-contracts/compare/v0.7.15...v0.8.0) (2023-11-20)
+
+
+### Bug Fixes
+
+* add gitignore ([cb22f28](https://github.com/rariblecom/protocol-contracts/commit/cb22f289ca98e233f86689a259e1092241d0077e))
+* remove generated files ([13f6ee7](https://github.com/rariblecom/protocol-contracts/commit/13f6ee72deab87f92c65114579d33fbfc0472de4))
+
+
+
+
+
 ## [0.7.15](https://github.com/rariblecom/protocol-contracts/compare/v0.7.14-beta.3...v0.7.15) (2023-10-10)
 
 **Note:** Version bump only for package @rarible/exchange-interfaces